+++ Cesium3DTilesSelection/src/LayerJsonTerrainLoader.cpp
@@ -1,16 +1,57 @@
 #include "LayerJsonTerrainLoader.h"
 
+#include "Cesium3DTilesSelection/BoundingVolume.h"
+#include "Cesium3DTilesSelection/Tile.h"
+#include "Cesium3DTilesSelection/TileContent.h"
+#include "Cesium3DTilesSelection/TileLoadResult.h"
+#include "Cesium3DTilesSelection/TilesetContentLoader.h"
+#include "Cesium3DTilesSelection/TilesetExternals.h"
+#include "Cesium3DTilesSelection/TilesetOptions.h"
+#include "CesiumAsync/AsyncSystem.h"
+#include "CesiumAsync/Future.h"
+#include "CesiumAsync/HttpHeaders.h"
+#include "CesiumAsync/IAssetAccessor.h"
+#include "CesiumGeometry/Axis.h"
+#include "CesiumGeometry/QuadtreeTileID.h"
+#include "CesiumGeometry/QuadtreeTileRectangularRange.h"
+#include "CesiumGeometry/QuadtreeTilingScheme.h"
+#include "CesiumGeometry/Rectangle.h"
+#include "CesiumGeospatial/BoundingRegionWithLooseFittingHeights.h"
+#include "CesiumGeospatial/Ellipsoid.h"
+#include "CesiumGeospatial/GeographicProjection.h"
+#include "CesiumGeospatial/GlobeRectangle.h"
+#include "CesiumGeospatial/Projection.h"
+#include "CesiumGeospatial/WebMercatorProjection.h"
+#include "CesiumUtility/Assert.h"
+#include "CesiumUtility/ErrorList.h"
+#include "TilesetContentLoaderResult.h"
+
 #include <CesiumAsync/IAssetResponse.h>
 #include <CesiumGeospatial/calcQuadtreeMaxGeometricError.h>
 #include <CesiumGltfContent/GltfUtilities.h>
 #include <CesiumQuantizedMeshTerrain/QuantizedMeshLoader.h>
 #include <CesiumRasterOverlays/RasterOverlayUtilities.h>
 #include <CesiumUtility/JsonHelpers.h>
-#include <CesiumUtility/Log.h>
 #include <CesiumUtility/Uri.h>
 
+#include <fmt/core.h>
+#include <glm/ext/matrix_double4x4.hpp>
+#include <gsl/span>
 #include <libmorton/morton.h>
 #include <rapidjson/document.h>
+#include <spdlog/logger.h>
+#include <spdlog/spdlog.h>
+
+#include <algorithm>
+#include <cmath>
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <optional>
+#include <string>
+#include <utility>
+#include <variant>
+#include <vector>
 
 using namespace CesiumAsync;
 using namespace Cesium3DTilesSelection;
@@ -390,7 +431,7 @@ Future<LoadLayersResult> loadLayersRecursive(
                 loadLayersResult.errors.emplaceWarning(fmt::format(
                     "Error when parsing layer.json, error code {} at byte "
                     "offset {}",
-                    layerJson.GetParseError(),
+                    static_cast<int>(layerJson.GetParseError()),
                     layerJson.GetErrorOffset()));
                 return asyncSystem.createResolvedFuture(
                     std::move(loadLayersResult));
@@ -496,7 +537,7 @@ Future<LoadLayersResult> loadLayerJson(
     LoadLayersResult result;
     result.errors.emplaceError(fmt::format(
         "Error when parsing layer.json, error code {} at byte offset {}",
-        layerJson.GetParseError(),
+        static_cast<int>(layerJson.GetParseError()),
         layerJson.GetErrorOffset()));
     return asyncSystem.createResolvedFuture(std::move(result));
   }
