+++ Cesium3DTilesSelection/src/TilesetJsonLoader.cpp
@@ -1,7 +1,24 @@
 #include "TilesetJsonLoader.h"
 
+#include "Cesium3DTilesContent/GltfConverterResult.h"
+#include "Cesium3DTilesSelection/BoundingVolume.h"
+#include "Cesium3DTilesSelection/Tile.h"
+#include "Cesium3DTilesSelection/TileContent.h"
+#include "Cesium3DTilesSelection/TileLoadResult.h"
+#include "Cesium3DTilesSelection/TileRefine.h"
+#include "Cesium3DTilesSelection/TilesetContentLoader.h"
+#include "Cesium3DTilesSelection/TilesetExternals.h"
+#include "CesiumAsync/HttpHeaders.h"
+#include "CesiumAsync/IAssetAccessor.h"
+#include "CesiumAsync/IAssetRequest.h"
+#include "CesiumGeometry/Axis.h"
+#include "CesiumGeospatial/Ellipsoid.h"
+#include "CesiumGeospatial/S2CellID.h"
+#include "CesiumGltfReader/GltfReader.h"
+#include "CesiumUtility/ErrorList.h"
 #include "ImplicitOctreeLoader.h"
 #include "ImplicitQuadtreeLoader.h"
+#include "TilesetContentLoaderResult.h"
 #include "logTileLoadResult.h"
 
 #include <Cesium3DTilesContent/GltfConverters.h>
@@ -17,13 +34,29 @@
 #include <CesiumGeospatial/S2CellBoundingVolume.h>
 #include <CesiumUtility/Assert.h>
 #include <CesiumUtility/JsonHelpers.h>
-#include <CesiumUtility/Log.h>
 #include <CesiumUtility/Uri.h>
-#include <CesiumUtility/joinToString.h>
 
+#include <fmt/core.h>
+#include <glm/common.hpp>
+#include <glm/ext/matrix_double4x4.hpp>
+#include <glm/geometric.hpp>
+#include <gsl/span>
 #include <rapidjson/document.h>
+#include <rapidjson/rapidjson.h>
+#include <spdlog/logger.h>
+#include <spdlog/spdlog.h>
 
+#include <algorithm>
 #include <cctype>
+#include <cstddef>
+#include <cstdint>
+#include <cstring>
+#include <memory>
+#include <optional>
+#include <string>
+#include <utility>
+#include <variant>
+#include <vector>
 
 using namespace CesiumUtility;
 using namespace Cesium3DTilesContent;
@@ -697,7 +730,7 @@ TileLoadResult parseExternalTilesetInWorkerThread(
     SPDLOG_LOGGER_ERROR(
         pLogger,
         "Error when parsing tileset JSON, error code {} at byte offset {}",
-        tilesetJson.GetParseError(),
+        static_cast<int>(tilesetJson.GetParseError()),
         tilesetJson.GetErrorOffset());
     return TileLoadResult::createFailedResult(std::move(pCompletedRequest));
   }
@@ -802,7 +835,7 @@ TilesetJsonLoader::createLoader(
           result.errors.emplaceError(fmt::format(
               "Error when parsing tileset JSON, error code {} at byte offset "
               "{}",
-              tilesetJson.GetParseError(),
+              static_cast<int>(tilesetJson.GetParseError()),
               tilesetJson.GetErrorOffset()));
           return asyncSystem.createResolvedFuture(std::move(result));
         }
