+++ CesiumQuantizedMeshTerrain/src/QuantizedMeshLoader.cpp
@@ -1,4 +1,20 @@
-#include <CesiumAsync/IAssetResponse.h>
+#include "CesiumGeometry/QuadtreeTileID.h"
+#include "CesiumGeospatial/BoundingRegion.h"
+#include "CesiumGeospatial/Ellipsoid.h"
+#include "CesiumGltf/Accessor.h"
+#include "CesiumGltf/Buffer.h"
+#include "CesiumGltf/BufferView.h"
+#include "CesiumGltf/Image.h"
+#include "CesiumGltf/Material.h"
+#include "CesiumGltf/MaterialPBRMetallicRoughness.h"
+#include "CesiumGltf/Mesh.h"
+#include "CesiumGltf/MeshPrimitive.h"
+#include "CesiumGltf/Model.h"
+#include "CesiumGltf/Node.h"
+#include "CesiumGltf/Sampler.h"
+#include "CesiumGltf/Scene.h"
+#include "CesiumGltf/Texture.h"
+
 #include <CesiumGeometry/QuadtreeTileRectangularRange.h>
 #include <CesiumGeospatial/GlobeRectangle.h>
 #include <CesiumGeospatial/calcQuadtreeMaxGeometricError.h>
@@ -6,18 +22,28 @@
 #include <CesiumQuantizedMeshTerrain/QuantizedMeshLoader.h>
 #include <CesiumUtility/AttributeCompression.h>
 #include <CesiumUtility/JsonHelpers.h>
-#include <CesiumUtility/Log.h>
 #include <CesiumUtility/Math.h>
 #include <CesiumUtility/Tracing.h>
-#include <CesiumUtility/Uri.h>
 
+#include <fmt/core.h>
 #include <glm/common.hpp>
-#include <glm/mat4x4.hpp>
-#include <glm/vec3.hpp>
+#include <glm/ext/vector_double3.hpp>
+#include <glm/ext/vector_float3.hpp>
+#include <glm/geometric.hpp>
+#include <gsl/span>
+#include <rapidjson/document.h>
+#include <rapidjson/rapidjson.h>
 
 #include <algorithm>
 #include <cstddef>
+#include <cstdint>
+#include <cstring>
+#include <limits>
+#include <optional>
 #include <stdexcept>
+#include <string>
+#include <utility>
+#include <vector>
 
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
@@ -1202,7 +1228,7 @@ static QuantizedMeshMetadataResult processMetadata(
   if (metadata.HasParseError()) {
     result.errors.emplaceError(fmt::format(
         "Error when parsing metadata, error code {} at byte offset {}",
-        metadata.GetParseError(),
+        static_cast<int>(metadata.GetParseError()),
         metadata.GetErrorOffset()));
     return result;
   }
