+++ CesiumRasterOverlays/src/IonRasterOverlay.cpp
@@ -1,3 +1,8 @@
+#include "CesiumAsync/Future.h"
+#include "CesiumRasterOverlays/RasterOverlay.h"
+#include "CesiumUtility/CreditSystem.h"
+#include "CesiumUtility/IntrusivePointer.h"
+
 #include <CesiumAsync/IAssetAccessor.h>
 #include <CesiumAsync/IAssetResponse.h>
 #include <CesiumRasterOverlays/BingMapsRasterOverlay.h>
@@ -7,8 +12,19 @@
 #include <CesiumRasterOverlays/RasterOverlayTileProvider.h>
 #include <CesiumRasterOverlays/TileMapServiceRasterOverlay.h>
 #include <CesiumUtility/JsonHelpers.h>
-#include <CesiumUtility/Log.h>
 #include <CesiumUtility/Uri.h>
+
+#include <fmt/core.h>
+#include <nonstd/expected.hpp>
+#include <rapidjson/document.h>
+#include <spdlog/logger.h>
+
+#include <cstdint>
+#include <memory>
+#include <string>
+#include <unordered_map>
+#include <utility>
+#include <vector>
 
 using namespace CesiumAsync;
 using namespace CesiumUtility;
@@ -125,7 +141,7 @@ IonRasterOverlay::createTileProvider(
                   fmt::format(
                       "Error while parsing Cesium ion raster overlay response, "
                       "error code {} at byte offset {}",
-                      response.GetParseError(),
+                      static_cast<int>(response.GetParseError()),
                       response.GetErrorOffset())});
             }
